Fix code style issues (#412)
- adds final to fields that are immutable
- replaces length() checks with isEmpty() calls
- reduces code that made raw use of parameterized types
- replaces if cascades with switch statements
- inlines variables

+++ dkpro-jwpl-api/src/main/java/org/dkpro/jwpl/api/util/ApiUtilities.java
@@ -68,7 +68,7 @@ public class ApiUtilities
         }
 
         if (counter % (size / step) == 0) {
-            double progressPercent = counter * 100 / size;
+            double progressPercent = (double) (counter * 100) / size;
             progressPercent = 1 + Math.round(progressPercent * 100) / 100.0;
             if (mode.equals(ApiUtilities.ProgressInfoMode.TEXT)) {
                 logger.info(text + ": " + progressPercent + " - " + OS.getUsedMemory() + " MB");
